3.4 update

+++ test_cuda.py
@@ -2,7 +2,7 @@ import torch
 
 def check_device():
     USE_CPU  = True
-    dtype = torch.float32
+    dtype = torch.float64
     if USE_CPU and torch.cuda.is_available():
         device = torch.device('cuda')
     else:
